Started Summary Statistics
We started summarizing our data, we had to split the variables into two types (numeric, categorical). For the numeric variables, we calculated their means. Right now, we are working on counting the values for the categorical data (eg. how many males and females do we have in our data set)